Add New Relic distributed logging support
This change adds a new function (Logger.NewContext) to the Logger
interface, which allows wrapping a Logger with a new context.Context.
This is useful for logging with distributed tracing.

The log15 submodule was instrumented to take advantage of this, so
that the logs sent to New Relic now have the full tracing
information. #minor